% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biaxial_plot.R
\name{biaxial_plot}
\alias{biaxial_plot}
\title{Generate a biaxial dot plot with contouring of the SCE data}
\usage{
biaxial_plot(
  indata,
  assay = "scaled",
  marker_x,
  marker_y,
  clustering = NULL,
  colkey = NULL
)
}
\arguments{
\item{indata}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{the SCE assay to choose from.}

\item{marker_x}{marker to display on the x axis}

\item{marker_y}{marker to display on the y axis}

\item{clustering}{metadata slot to colour the plot by}

\item{colkey}{a list of colours matching the number of conditions found in clustering}
}
\value{
a \code{ggplot} object.
}
\description{
Generate a biaxial dot plot with contouring of the SCE data
}
\examples{
# Download complete ImmunoCluster SCE object from zenodo
sce_gvhd = readRDS(url("https://zenodo.org/record/3801882/files/sce_gvhd.rds"))

biaxial_plot(sce_gvhd, marker_x = "CD8a", marker_y = "CD4")


}
\author{
James Opzoomer \email{james.opzoomer@kcl.ac.uk}
}
