% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsplot.R
\name{mdsplot}
\alias{mdsplot}
\title{Create sample-level MDS plot}
\usage{
mdsplot(
  data,
  grouping = "group",
  feature = NULL,
  markers = NULL,
  colkey = NULL,
  legendPosition = "right",
  legendLabSize = 12,
  legendIconSize = 5,
  xlim = NULL,
  ylim = NULL,
  labSize = 3,
  labhjust = 1.5,
  labvjust = 0,
  drawConnectors = TRUE,
  widthConnectors = 0.5,
  colConnectors = "black",
  xlab = dimColnames[1],
  xlabAngle = 0,
  xlabhjust = 0.5,
  xlabvjust = 0.5,
  ylab = dimColnames[2],
  ylabAngle = 0,
  ylabhjust = 0.5,
  ylabvjust = 0.5,
  axisLabSize = 16,
  title = "Metadata plot",
  subtitle = "",
  caption = "MDS Plot",
  titleLabSize = 16,
  subtitleLabSize = 12,
  captionLabSize = 12,
  hline = NULL,
  hlineType = "longdash",
  hlineCol = "black",
  hlineWidth = 0.4,
  vline = NULL,
  vlineType = "longdash",
  vlineCol = "black",
  vlineWidth = 0.4,
  gridlines.major = TRUE,
  gridlines.minor = TRUE,
  borderWidth = 0.8,
  borderColour = "black"
)
}
\arguments{
\item{data}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{grouping}{a character string representaing which \code{metadata}
slot to group the individual samples by.}

\item{feature}{a character string represnting a feature to label the heatmap with.}

\item{markers}{a character string specifying which markers to include.}

\item{colkey}{a vector of grouping ids the colour strings to indicate the feature labelling.}

\item{legendPosition}{a string for \code{ggplot2} legend position.}

\item{legendLabSize}{\code{ggplot2} legend label size}

\item{legendIconSize}{\code{ggplot2} legend icon size}

\item{xlim}{\code{ggplot2} x limit specification}

\item{ylim}{\code{ggplot2} y limit specification}

\item{xlabAngle}{specify the x lab angle}

\item{xlabhjust}{Specify the horizontal justification of the x axis (0 = left jsutified, 1 = right justified)}

\item{xlabvjust}{Specify the vertical justification of the x axis (0 = left jsutified, 1 = right justified)}

\item{ylab}{specify the y axis label}

\item{ylabAngle}{specify the y lab angle}

\item{ylabhjust}{Specify the horizontal justification of the y axis (0 = left jsutified, 1 = right justified)}

\item{ylabvjust}{Specify the vertical justification of the y axis (0 = left jsutified, 1 = right justified)}

\item{axisLabSize}{specify the axis label size}

\item{title}{specify the title text of the plot}

\item{subtitle}{specify the subtitle text of the plot}

\item{caption}{specify the caption text of the plot}

\item{titleLabSize}{specify the title text size}

\item{subtitleLabSize}{specify the subtitle text size}

\item{captionLabSize}{specify the caption text size}

\item{stripLabSize}{specify the facet labels of a plot}
}
\value{
a \code{ggplot} object.
}
\description{
Multi-Dimensional Scaling (MDS) plot computed across
median marker expression in each sample or metadata grouping.
}
\examples{
# Download complete ImmunoCluster SCE object from zenodo
sce_gvhd = readRDS(url("https://zenodo.org/record/3801882/files/sce_gvhd.rds"))

# Generate sample level MDS plot with all markers
mdsplot(sce_gvhd, feature = "condition", colkey = c(None = 'royalblue', GvHD = 'red2'))

}
\author{
James Opzoomer
}
