% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performTSNE.R
\name{performTSNE}
\alias{performTSNE}
\title{Perform TSNE dimensionality reduction on SingleCellExperiment object}
\usage{
performTSNE(
  indata,
  assay = "scaled",
  downsample = NULL,
  downsample_grouping = "group",
  reducedDim = NULL,
  dims = seq_len(20),
  newDimName = NULL,
  useMarkers = NULL,
  perplexity = 50,
  theta = 0.5,
  max_iter = 1000
)
}
\arguments{
\item{indata}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{the SCE assay to choose from.}

\item{downsample}{a numeric of the number of cells to downsample by.}

\item{downsample_grouping}{a character string indicating which metadata grouping to downsample by.}

\item{reducedDim}{a character vector represting a dimensionality reduction slot from the SCE to be selected as input data.}

\item{dims}{numeric of reduced dimensions to select from the reducedDim}

\item{newDimName}{character string of the name for the resulting TSNE dimensionality reduction in the SCE object.}

\item{useMarkers}{chacter vector of markers to use as input for the TSNE run.}

\item{perplexity}{numeric of the perplexity for the TSNE algorithm param.}

\item{theta}{numeric of the theta for the TSNE algorithm param.}

\item{max_iter}{numeric of the maximum iterations to perform for the TSNE algorithm param.}
}
\value{
a \code{SingleCellExperiment} object.
}
\description{
Perform TSNE dimensionality reduction on SingleCellExperiment object
}
\examples{
# Download complete ImmunoCluster SCE object from zenodo
sce_gvhd = readRDS(url("https://zenodo.org/record/3801882/files/sce_gvhd.rds"))

# Run tSNE and store in sce object
sce_gvhd = performTSNE(sce_gvhd)

}
\author{
Kevin Blighe, James Opzoomer \email{james.opzoomer@kcl.ac.uk}
}
