% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performUMAP.R
\name{performUMAP}
\alias{performUMAP}
\title{Perform UMAP dimensionality reduction on SingleCellExperiment object}
\usage{
performUMAP(
  indata,
  assay = "scaled",
  downsample = NULL,
  downsample_grouping = "group",
  reducedDim = NULL,
  dims = seq_len(20),
  newDimName = NULL,
  useMarkers = NULL
)
}
\arguments{
\item{indata}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{the SCE assay to choose from.}

\item{downsample}{a numeric of the number of cells to downsample by.}

\item{downsample_grouping}{a character string indicating which metadata grouping to downsample by.}

\item{reducedDim}{a character vector represting a dimensionality reduction slot from the SCE to be selected as input data.}

\item{dims}{numeric of reduced dimensions to select from the reducedDim}

\item{newDimName}{character string of the name for the resulting UMAP dimensionality reduction in the SCE object.}

\item{useMarkers}{chacter vector of markers to use as input for the UMAP run.}
}
\value{
a \code{SingleCellExperiment} object.
}
\description{
Perform UMAP dimensionality reduction on SingleCellExperiment object
}
\examples{
# Download complete ImmunoCluster SCE object from zenodo
sce_gvhd = readRDS(url("https://zenodo.org/record/3801882/files/sce_gvhd.rds"))

# Run UMAP and store in sce object
sce_gvhd = performUMAP(sce_gvhd)

}
\author{
Kevin Blighe, James Opzoomer \email{james.opzoomer@kcl.ac.uk}
}
