% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbundance.R
\name{plotAbundance}
\alias{plotAbundance}
\title{Create a ggplot object of sce cluster proportion abundances grouped by sample and metadata condition}
\usage{
plotAbundance(
  indata,
  graph_type = "box",
  group = "group",
  clusters = unique(metadata(indata)[["cell_annotation"]]),
  clusterAssign = "cell_annotation",
  feature = "condition",
  colkey = NULL,
  legendPosition = "right",
  pointSize = 1,
  legendLabSize = 12,
  legendIconSize = 5,
  legendKeyHeight = 2.5,
  xlim = NULL,
  ylim = NULL,
  xlab = "Population",
  xlabAngle = 90,
  xlabhjust = 0.5,
  xlabvjust = 0.5,
  ylab = "\% of cells",
  ylabAngle = 0,
  ylabhjust = 0.5,
  ylabvjust = 0.5,
  axisLabSize = 16,
  stripLabSize = 16,
  title = "Marker expression per cluster",
  subtitle = "",
  caption = "",
  titleLabSize = 16,
  subtitleLabSize = 12,
  captionLabSize = 12,
  borderWidth = 0.8,
  borderColour = "black"
)
}
\arguments{
\item{indata}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}}}

\item{graph_type}{Whether to create a boxplot with 'box' or stacked barchart with 'bar'}

\item{group}{The metadata slot representing sample_id}

\item{clusters}{The specific clusters to plot within the clustering metatdata slot}

\item{clusterAssign}{the clustering metadata or other metadata slot by which to group the samples specified in 'group'}

\item{feature}{A feature representing a metadata slot by which to split the plots}

\item{colkey}{A colour key by which to manually specify plotting colours}

\item{legendPosition}{Where to position the legend}

\item{pointSize}{Dotplot pointsize}

\item{legendLabSize}{legend label size}

\item{legendIconSize}{legend key icon size}

\item{legendKeyHeight}{legend key height}

\item{xlim}{specify the x limit}

\item{ylim}{specify the y limit}

\item{xlab}{specify the x axis label}

\item{xlabAngle}{specify the x lab angle}

\item{xlabhjust}{Specify the horizontal justification of the x axis (0 = left jsutified, 1 = right justified)}

\item{xlabvjust}{Specify the vertical justification of the x axis (0 = left jsutified, 1 = right justified)}

\item{ylab}{specify the y axis label}

\item{ylabAngle}{specify the y lab angle}

\item{ylabhjust}{Specify the horizontal justification of the y axis (0 = left jsutified, 1 = right justified)}

\item{ylabvjust}{Specify the vertical justification of the y axis (0 = left jsutified, 1 = right justified)}

\item{axisLabSize}{specify the axis label size}

\item{stripLabSize}{specify the facet labels of a plot}

\item{title}{specify the title text of the plot}

\item{subtitle}{specify the subtitle text of the plot}

\item{caption}{specify the caption text of the plot}

\item{titleLabSize}{specify the title text size}

\item{subtitleLabSize}{specify the subtitle text size}

\item{captionLabSize}{specify the caption text size}
}
\description{
Create a ggplot object of sce cluster proportion abundances grouped by sample and metadata condition
}
