% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFlowSOM.R
\name{runFlowSOM}
\alias{runFlowSOM}
\title{Perform combined FlowSOM clustering followed by consensus clustering as outlined in Nowicka et al. (2017) F1000Research.}
\usage{
runFlowSOM(
  sct,
  k = 30,
  markers = NULL,
  som_x = 10,
  som_y = 10,
  set_seed = 4321,
  scale = FALSE
)
}
\arguments{
\item{sct}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{k}{a numeric representing the desired maximum number of clusters to group cells into by consensus clustering, the funtion will return all clustering k values from 2-k and these will be stored in the metadata table.}

\item{markers}{a vector of strings representing the markers to cluster on.}

\item{som_x}{a numeric indicating the x dimension of the som grid, see FLowSOM documentation.}

\item{som_y}{a numeric indicating the y dimension of the som grid.}

\item{set_seed}{a numeric to set the random seed for run reproducibility.}

\item{scale}{a boolean indicating whether to scale the input data by marker, a parameter to the flowSOM ReadInput command.}
}
\value{
a \code{SingleCellExperiment} object with new metadata slots representing consensus clustering of the flowSOM codes from 2 to k.
}
\description{
Perform combined FlowSOM clustering followed by consensus clustering as outlined in Nowicka et al. (2017) F1000Research.
}
\examples{
# Download complete ImmunoCluster SCE object from zenodo
sce_gvhd = readRDS(url("https://zenodo.org/record/3801882/files/sce_gvhd.rds"))

# Run FlowSOM and store in sce object to clustering k=
sce_gvhd = runFlowSOM(sce_gvhd, k = 10, markers = clustering_markers, som_x = 10, som_y = 10)

}
\author{
James Opzoomer \email{james.opzoomer@kcl.ac.uk}
}
