% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhenograph.R
\name{runPhenograph}
\alias{runPhenograph}
\title{run Rphenograph implementation from "JinmiaoChenLab/Rphenograph" clustering on sce expression data}
\usage{
runPhenograph(sce, k = 30, markers = NULL, set_seed = 4321)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{k}{a number of desired clusters to cluster cell data into.}

\item{markers}{a vector of selected markers, representing sce colnames to use for clustering}

\item{set_seed}{set the random seed for clustering reproducibility}
}
\value{
a \code{SingleCellExperiment} object with a new metadata slot with cluster identities.
}
\description{
run Rphenograph implementation from "JinmiaoChenLab/Rphenograph" clustering on sce expression data
}
