% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClusterIdents.R
\name{setClusterIdents}
\alias{setClusterIdents}
\title{Relabel clustering identities in sce metadata.}
\usage{
setClusterIdents(sce, orig.ident, new.ident, clustering = "phenograph")
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{orig.ident}{a vector of original cluster names present.}

\item{new.ident}{a vector of the same length with the new cluster identities to map orig.ident to.}

\item{clustering}{the name of the clustering metadata slot to map the orig.ident to.}
}
\value{
an sce with a new metadata slot called 'cell_annotation' containing the new cluster mappings.
}
\description{
Relabel clustering identities in sce metadata.
}
