% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test_clust.R
\name{stat_test_clust}
\alias{stat_test_clust}
\title{Perform statistical test on cluster abundances as a proportion of total cells within a sample across conditions.
Generates p adjusted values with "BH" p.adjust method.}
\usage{
stat_test_clust(
  sct,
  group = "group",
  clustering = "cell_annotation",
  feature = "condition",
  test = "wilcox",
  var_equal = TRUE
)
}
\arguments{
\item{sct}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{group}{the sample_id to generate the calculate the cluster abundance proportions from.}

\item{clustering}{the clustering to the cluster abundance proportions from.}

\item{feature}{the metadata condition to test the cluster abundance proportions between.}

\item{test}{either 'wilcox' calls pairwise.wilcox.test or 't_test' calls pairwise.t.test.}

\item{var_equal}{sets the pool.sd parameter to apply welch's correction as a parameter of pairwise.t.test}
}
\value{
a dataframe containing the aubundance proportions per group of each cluster
along with pval and BH adjusted pvalue.
}
\description{
Perform statistical test on cluster abundances as a proportion of total cells within a sample across conditions.
Generates p adjusted values with "BH" p.adjust method.
}
