% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test_expression.R
\name{stat_test_expression}
\alias{stat_test_expression}
\title{Perform statistical test on median expression values across clusters and conditions.
Generates p adjusted values with "BH" p.adjust method.}
\usage{
stat_test_expression(
  indata,
  assay = "scaled",
  grouping = "group",
  feature = "condition",
  clusterAssign = "cell_annotation",
  test = "wilcox",
  var_equal = TRUE
)
}
\arguments{
\item{indata}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{assay}{the assay to select from the sce.}

\item{grouping}{the sample_ids to generate the median expression from.}

\item{feature}{the feature to test the median expression data between.}

\item{clusterAssign}{the clustering to generate the median expression from.}

\item{test}{either 'wilcox' calls pairwise.wilcox.test or 't_test' calls pairwise.t.test.}

\item{var_equal}{sets the pool.sd parameter to apply welch's correction as a parameter of pairwise.t.test}
}
\value{
a dataframe containing the median marker expression per cluster-marker and sample
along with pval and BH adjusted pvalue.
}
\description{
Perform statistical test on median expression values across clusters and conditions.
Generates p adjusted values with "BH" p.adjust method.
}
