% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_sce_metadata.R
\name{subset_sce_metadata}
\alias{subset_sce_metadata}
\title{Subset the sce object based on a column of the metadata table}
\usage{
subset_sce_metadata(sce, ...)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{...}{a conditional statement to subset the SCE by metadata slot.}
}
\value{
a subsetted sce \code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}
\description{
Subset the sce object based on a column of the metadata table
}
\examples{

 # Subset the sce_gvhd object based on the 'cell_annotation' label in the metadata table
 sce_cd4 = subset_sce_metadata(sce_gvhd, cell_annotation \%in\% c("CD4+ T Cells", "Basophils"))

}
