% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Generate a volcano plot using statistical testing output}
\usage{
volcanoPlot(pval_df, p_val = "padj", threshold = 0.1)
}
\arguments{
\item{pval_df}{a dataframe that contains a column name "logfc" and a statistical testing output p values.}

\item{p_val}{a string identifying the column name from pval_df to plot as p values}

\item{threshold}{a numeric representing the p_value threshold on which to colour points on the volcano plot}
}
\value{
a \code{ggplot} object.
}
\description{
Generate a volcano plot using statistical testing output
}
\examples{

}
\author{
James Opzoomer \email{james.opzoomer@kcl.ac.uk}
}
