% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sce_to_fcs.R
\name{write_sce_to_fcs}
\alias{write_sce_to_fcs}
\title{Export an SCE to.fcs format}
\usage{
write_sce_to_fcs(sce, filename = "export.fcs", clustering = NULL)
}
\arguments{
\item{sce}{a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{filename}{the .fcs file name to export}

\item{clustering}{metadata clustering slots to export as parameters in the .fcs file}
}
\value{
a .fcs file
}
\description{
Takes as input a SingleCellExperiment object and writes
the data contained into an .fcs file
}
\examples{
# Download complete ImmunoCluster SCE object from zenodo
sce_gvhd = readRDS(url("https://zenodo.org/record/3801882/files/sce_gvhd.rds"))

# write sct to fcs
write_sce_to_fcs(sce_gvhd)

}
\author{
James Opzoomer \email{james.opzoomer@kcl.ac.uk}
}
